Added some docu, minor bugfix, removed FIXMEs
Bugfix: sparse should have been called with dimensions in case there are
rows/cols of all zeros at the end.  This probably doesn't occur, but
better be safe.  Also should be slightly more efficient.

I decided not to do the find() for params in functions_setup because it
would have introduced ugly dependencies between matlab_encoder and
MatlabCodegen.  Eric's manual find indexing is good and self contained
as is.